requirements.txt
pandas==1.5.1
pyodbc==5.0.1
python-dotenv==1.0.0
python_dateutil==2.8.2
requests==2.28.1
